requirements.txt
_libgcc_mutex==0.1
_openmp_mutex==5.1
accelerate==1.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.10.5
aiosignal==1.3.2
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
arrow-cpp==17.0.0
asttokens==3.0.0
async-lru==2.0.4
attrs==24.3.0
aws-c-auth==0.6.19
aws-c-cal==0.5.20
aws-c-common==0.8.5
aws-c-compression==0.2.16
aws-c-event-stream==0.2.15
aws-c-http==0.6.25
aws-c-io==0.13.10
aws-c-mqtt==0.7.13
aws-c-s3==0.1.51
aws-c-sdkutils==0.1.6
aws-checksums==0.1.13
aws-crt-cpp==0.18.16
aws-sdk-cpp==1.11.212
babel==2.16.0
beautifulsoup4==4.12.3
blas==1.0
bleach==6.2.0
bleach-with-css==6.2.0
boost-cpp==1.82.0
bottleneck==1.4.2
brotli==1.0.9
brotli-python==1.0.9
bzip2==1.0.8
c-ares==1.19.1
ca-certificates==2024.12.14
cached-property==1.5.2
cached_property==1.5.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cuda-cudart==12.1.105
cuda-cupti==12.1.105
cuda-libraries==12.1.0
cuda-nvrtc==12.1.105
cuda-nvtx==12.1.105
cuda-opencl==12.6.77
cuda-runtime==12.1.0
cuda-version==12.6
cycler==0.11.0
cyrus-sasl==2.1.28
datasets==3.2.0
dbus==1.13.18
debugpy==1.8.11
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
entrypoints==0.4
exceptiongroup==1.2.2
executing==2.1.0
expat==2.6.4
ffmpeg==4.3
filelock==3.13.1
font-ttf-dejavu-sans-mono==2.37
font-ttf-inconsolata==3.000
font-ttf-source-code-pro==2.038
font-ttf-ubuntu==0.83
fontconfig==2.14.1
fonts-conda-forge==1
fonttools==4.51.0
fqdn==1.5.1
freetype==2.12.1
frozenlist==1.5.0
fsspec==2024.9.0
gflags==2.2.2
giflib==5.2.2
glib==2.78.4
glib-tools==2.78.4
glog==0.5.0
gmp==6.2.1
gnutls==3.6.15
gst-plugins-base==1.14.1
gstreamer==1.14.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.28.1
huggingface_hub==0.27.1
hyperframe==6.0.1
icu==73.1
idna==3.7
importlib-metadata==8.5.0
importlib_resources==6.5.2
intel-openmp==2023.1.0
ipykernel==6.29.5
ipython==8.31.0
isoduration==20.11.0
jedi==0.19.2
jinja2==3.1.4
joblib==1.4.2
jpeg==9e
json5==0.10.0
jsonpointer==2.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jsonschema-with-format-nongpl==4.23.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_events==0.11.0
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kaleido-core==0.2.1
kiwisolver==1.4.4
krb5==1.20.1
lame==3.100
lcms2==2.16
ld_impl_linux-64==2.40
lerc==4.0.0
libabseil==20240116.2
libboost==1.82.0
libbrotlicommon==1.0.9
libbrotlidec==1.0.9
libbrotlienc==1.0.9
libclang==14.0.6
libclang13==14.0.6
libcublas==12.1.0.26
libcufft==11.0.2.4
libcufile==1.11.1.6
libcups==2.4.2
libcurand==10.3.7.77
libcurl==8.11.1
libcusolver==11.4.4.55
libcusparse==12.0.2.55
libdeflate==1.22
libedit==3.1.20230828
libev==4.33
libevent==2.1.12
libffi==3.4.4
libgcc-ng==11.2.0
libgfortran-ng==11.2.0
libgfortran5==11.2.0
libglib==2.78.4
libgomp==11.2.0
libgrpc==1.62.2
libiconv==1.16
libidn2==2.3.4
libjpeg-turbo==2.0.0
libllvm14==14.0.6
libnghttp2==1.57.0
libnpp==12.0.2.50
libnvjitlink==12.1.105
libnvjpeg==12.1.1.14
libpng==1.6.39
libpq==17.2
libprotobuf==4.25.3
libsodium==1.0.18
libssh2==1.11.1
libstdcxx-ng==11.2.0
libtasn1==4.19.0
libthrift==0.15.0
libtiff==4.5.1
libunistring==0.9.10
libuuid==1.41.5
libwebp==1.3.2
libwebp-base==1.3.2
libxcb==1.15
libxkbcommon==1.0.1
libxml2==2.13.5
llvm-openmp==14.0.6
lz4-c==1.9.4
markupsafe==2.1.3
mathjax==2.7.7
matplotlib==3.10.0
matplotlib-base==3.10.0
matplotlib-inline==0.1.7
mistune==3.1.0
mkl==2023.1.0
mkl-service==2.4.0
mkl_fft==1.3.11
mkl_random==1.2.8
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.15
mysql==8.4.0
nbclient==0.10.2
nbconvert-core==7.16.5
nbformat==5.10.4
ncurses==6.4
nest-asyncio==1.6.0
nettle==3.7.3
networkx==3.2.1
notebook-shim==0.2.4
nspr==4.35
nss==3.89.1
numexpr==2.10.1
numpy==1.26.4
numpy-base==1.26.4
openh264==2.1.1
openjpeg==2.5.2
openldap==2.6.4
openssl==3.0.15
orc==2.0.1
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.0
parso==0.8.4
pcre2==10.42
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.0.0
pip==24.2
pkgutil-resolve-name==1.3.10
platformdirs==4.3.6
plotly==5.24.1
ply==3.11
prometheus_client==0.21.1
prompt-toolkit==3.0.48
psutil==5.9.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pycparser==2.22
pygments==2.19.1
pyparsing==3.2.0
pyqt==5.15.10
pyqt5-sip==12.13.0
pysocks==1.7.1
python==3.12.8
python-dateutil==2.9.0.post0
python-fastjsonschema==2.21.1
python-json-logger==2.0.7
python-kaleido==0.2.1
python-tzdata==2024.2
python-xxhash==2.0.2
pytorch==2.5.1
pytorch-cuda==12.1
pytorch-mutex==1.0
pytz==2024.2
pyyaml==6.0.2
pyzmq==26.2.0
qt-main==5.15.2
re2==2022.04.01
readline==8.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.22.3
s2n==1.3.27
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
send2trash==1.8.3
sentence-transformers==3.3.1
setuptools==75.1.0
sip==6.7.12
six==1.17.0
snappy==1.2.1
sniffio==1.3.1
soupsieve==2.5
sqlite==3.45.3
stack_data==0.6.3
sympy==1.13.3
tbb==2021.8.0
tenacity==9.0.0
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.4.0
tk==8.6.14
tokenizers==0.20.1
tomli==2.2.1
torchaudio==2.5.1
torchtriton==3.1.0
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.46.1
types-python-dateutil==2.9.0.20241206
typing-extensions==4.12.2
typing_extensions==4.12.2
typing_utils==0.1.0
tzdata==2024b
unicodedata2==15.1.0
uri-template==1.3.0
urllib3==2.2.3
utf8proc==2.6.1
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
wheel==0.44.0
xxhash==0.8.0
xz==5.4.6
yaml==0.2.5
yarl==1.11.0
zeromq==4.3.5
zipp==3.21.0
zlib==1.2.13
zstd==1.5.6
